Module Bap_mips_target

Declares some of the MIPS targets.

More targets could be added without modifying this module. It is advised to inherit from one of the targets defined here to include the newly defined target into the mips family, though it is not strictly required.

val parent : Bap_core_theory.Theory.Target.t

parent it the parent of all MIPS.

No properties are set.

The MIPS 32 family

val mips32bi : Bap_core_theory.Theory.Target.t

The bi-endian MIPS32 (the parent)

val mips32eb : Bap_core_theory.Theory.Target.t

The bi-endian MIPS32 (the parent)

The big endian MIPS32

val mips32le : Bap_core_theory.Theory.Target.t

The big endian MIPS32

The little endian MIPS32

The MIPS 64 family

val mips64bi : Bap_core_theory.Theory.Target.t

The bi-endian MIPS64 (the parent)

val mips64eb : Bap_core_theory.Theory.Target.t

The bi-endian MIPS64 (the parent)

The big endian MIPS64

val mips64le : Bap_core_theory.Theory.Target.t

The big endian MIPS64

The little endian MIPS64

MIPS encodings

val llvm_mips32 : Bap_core_theory.Theory.language

val llvm_mips64 : Bap_core_theory.Theory.language

val load : ?backend:string -> unit -> unit

load () loads the knowledge base rules for the MIPS targets.

This includes parsing the loader output and enabling backward compatibility with the old Arch.t representation.